python 3.6 compat

+++ plotRFs.py
@@ -7,7 +7,7 @@ np.set_printoptions(precision=4, suppress=True)
 w = np.load('w.npy')
 NBCELLS = w.shape[0]
 SIDE = np.ceil(np.sqrt(NBCELLS )).astype(int)
-RFSIZE = np.sqrt(w.shape[1])
+RFSIZE = np.sqrt(w.shape[1]).astype(int)
 
 fig = plt.figure()
 
@@ -18,6 +18,3 @@ for nn in range(NBCELLS):
 
 plt.draw()
 plt.savefig('RFs.png', bbox_inches='tight')
-
-
-
